break up into header and code files

+++ lib/HT16K33Disp/HT16K33Disp.cpp
@@ -3,20 +3,20 @@
 #include <Arduino.h>
 #include "HT16K33Disp.h"
 
-HT16K33Disp::HT16K33Disp(int address, int num_displays){
+HT16K33Disp::HT16K33Disp(byte address, byte num_displays){
     set_address(address, num_displays);
     _loop_running = false;
 }
 
-void HT16K33Disp::set_address(int address, int num_displays){
+void HT16K33Disp::set_address(byte address, byte num_displays){
     _address = address;
     _num_displays = num_displays;
     _num_digits = _num_displays * NUM_DIGITS_PER_DISPLAY;
 }
 
 // point to an array of ints specifying brightness levels per display
-void HT16K33Disp::init(const int *brightLevels){
-    for(int i = 0; i < _num_displays; i++){
+void HT16K33Disp::init(const byte *brightLevels){
+    for(byte i = 0; i < _num_displays; i++){
         Wire.beginTransmission(_address + i);
         Wire.write(0x21);               //normal operation mode
         Wire.endTransmission(false);
@@ -30,7 +30,7 @@ void HT16K33Disp::init(const int *brightLevels){
     }
 }
 
-void HT16K33Disp::write(int digit, unsigned int data){
+void HT16K33Disp::write(byte digit, unsigned int data){
     int display = digit / NUM_DIGITS_PER_DISPLAY;
     digit -= (display * NUM_DIGITS_PER_DISPLAY);
     Wire.beginTransmission(_address + display);
@@ -41,12 +41,12 @@ void HT16K33Disp::write(int digit, unsigned int data){
 }
 
 void HT16K33Disp::segments_test(){
-    for(int i = 0; i < _num_digits; i++)
+    for(byte i = 0; i < _num_digits; i++)
         write(i, (uint16_t) -1);
 }
 
 void HT16K33Disp::clear(){
-    for(int i = 0; i < _num_digits; i++)
+    for(byte i = 0; i < _num_digits; i++)
         write(i, 0);
 }
 
@@ -68,7 +68,7 @@ int HT16K33Disp::string_length(const char * string){
 }
 
 void HT16K33Disp::show_string(const char * string, bool pad_blanks, bool right_justify){
-    int i = 0;
+    byte i = 0;
     if(right_justify)
     {
         char diff = _num_digits - string_length(string);
@@ -83,7 +83,7 @@ void HT16K33Disp::show_string(const char * string, bool pad_blanks, bool right_j
             i = diff;
     }
 
-    for(int j = i; j < _num_digits; j++)
+    for(byte j = i; j < _num_digits; j++)
     {
         if(*string == 0)
         {
@@ -108,7 +108,7 @@ void HT16K33Disp::show_string(const char * string, bool pad_blanks, bool right_j
 }
 
 void HT16K33Disp::simple_show_string(const char * string){
-    for(int i = 0; i < _num_digits; i++){
+    for(byte i = 0; i < _num_digits; i++){
         if(*string == 0)
             break;
         if(*(string + 1) == '.')
